Support Cassandra 5 and BTI SSTable format

+++ cassandra-analytics-core/src/test/java/org/apache/cassandra/spark/reader/CassandraBridgeUtilTests.java
@@ -164,7 +164,10 @@ public class CassandraBridgeUtilTests
                 }
 
                 // delete Summary.db file and check we can read the Index.db file too
-                Files.deleteIfExists(summaryDbFile);
+                if (summaryDbFile != null)
+                {
+                    Files.deleteIfExists(summaryDbFile);
+                }
             }
         });
     }
